Handling better stbl parameters
Several commands use the stbl parameter to select the DA Setup table.
I was using a default value as function argument. Now the default is
to use the Initium struct field `stbl`. If this value should be changed,
there is function `dtcsetstbl!(dev, stbl)` that sets the new (global) value.

For this to work, I didn't remove the stbl arguments from the functions.
Instead I set the default argument value -1. If stbl==-1, use the Initium
global field stbl.