Significant changes: a Initium device deals with a single stbl only
I simplified many things. But the basic idea is that if you want to use
more than a single stbl, you just create a new Initium, based on the
original that has deals with the new stbl.

Things appear to be working but I really need to stress test this
and document it well!