Now interrupting daq with Ctrl-C appears to be working
I don't know if it was triggered by all the Ctrl-C I was pressing but
a weird problem that we have sometimes experienced with the Initium, reading
less frames than requested has finally appeared  while developing the julia
code. The funny thing is that now, I was able to inspect the output exactly.
It just reads less frames than requested. Response type 04 (normal)
and response code 102, the same as when there is no problem! Weird...